+++ tests/test_serializer.py
@@ -2,14 +2,14 @@
 import json
 # Project imports
 from ocxtools.parser.parser import OcxParser
-from ocxtools.serializer.serializer import Serializer
+from ocxtools.serializer.serializer import OcxSerializer
 
 
 def test_serialize_json(shared_datadir, data_regression):
     parser = OcxParser()
     model = shared_datadir / "m1.3Docx"
     ocxxml = parser.parse(str(model.resolve()))
-    serializer = Serializer(ocxxml)
+    serializer = OcxSerializer(ocxxml)
     result = serializer.serialize_json()
     data_regression.check(json.loads(result))
 
@@ -18,6 +18,6 @@ def test_serialize_xml(shared_datadir):
     parser = OcxParser()
     model = shared_datadir / "m1.3Docx"
     ocxxml = parser.parse(str(model.resolve()))
-    serializer = Serializer(ocxxml)
+    serializer = OcxSerializer(ocxxml)
     result = serializer.serialize_xml()
     assert '?xml version' in result