+++ CHANGELOG.md
@@ -1,20 +1,29 @@
 # ``ocxtools``: Changelog
 
 
-## 1.2.1 2024.01.08
+## 1.3.0 - 2024.01.012
+
+### Changes
+ - Add support for docker validator v3.0.0b6
+ - Docker run will always pull :latest image as default
+ - Add saving of detail validation report to csv file
+
+[Release tag](https://github.com/OCXStandard/ocxtools/tree/v1.3.0)
+
+## 1.2.1 - 2024.01.08
 
 ### Changes
 Fix missing printout of stdout when running a sub process.
 
 
-## 1.2.0 2024.01.08
+## 1.2.0 - 2024.01.08
 
 ### Changes
 Add a docker UI validation command to the docker CLI.
 
 [Release tag](https://github.com/OCXStandard/ocxtools/tree/v1.2.0)
 
-## 1.1.0 2024.01.05
+## 1.1.0 - 2024.01.05
 
 ### 1st release